MNSTR-5245: Make the JS performance metrics opt in
