+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/w3c/harvesters/Common.kt
@@ -1,0 +1,11 @@
+package com.atlassian.performance.tools.jiraactions.api.w3c.harvesters
+
+import java.time.Duration
+
+internal fun parseTimestamp(
+    timestamp: Any?
+): Duration = when (timestamp) {
+    is Long -> Duration.ofMillis(timestamp)
+    is Double -> Duration.ofMillis(timestamp.toLong())
+    else -> throw Exception("Cannot parse timestamp from $timestamp")
+}
